% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_script.R
\name{find_script}
\alias{find_script}
\title{Internal Function for pathing of the building blocks for the template}
\usage{
find_script(script_name, package = "AnalysisMEP")
}
\arguments{
\item{script_name}{specifies which script is called.}

\item{package}{is the name of the package where the buildingblocks are included ("RAnalysisMEP").}
}
\value{
the path of the script_name
}
\description{
This is an internal function which will find the correct path for the buildingblocks.
}
\details{
This function adapts the code from the internal function find_template from the `usethis::use_template` function.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
